Flyttat om filer och modulariserat inom server
