--- server/src/example_tests_src.ts
@@ -1,10 +1,0 @@
-export function sum(a: number, b: number): number {
-    return a + b;
-}
-
-
-export function fetchData(callback: (data: string) => void): void {
-    setTimeout(() => {
-      callback('data');
-    }, 1000);
-}
